Extract Appointment Results to component
