+++ src/reports/AppointmentResourcePlanning/AppointmentSummary.tsx
@@ -1,5 +1,5 @@
 import React, {ReactElement} from "react";
-import {AppointmentResourcePlanningSummaryReportData} from "../interfaces";
+import {AppointmentResourcePlanningSummaryReportData} from "../../interfaces";
 import {ONSPanel} from "blaise-design-system-react-components";
 
 interface Props {
